pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kaggle-s3-lake"
version = "0.1.0"
description = "Download a public dataset and publish it locally as Parquet, Delta, and Iceberg"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Data Platform Team" }]
dependencies = [
  "certifi>=2024.0.0",
  "deltalake>=0.18",
  "pandas>=2.2",
  "pyarrow>=15.0",
  "pyiceberg[sql-sqlite]>=0.8.0",
  "python-dotenv>=1.0",
  "tabulate>=0.9"
]

[project.scripts]
kaggle-s3-lake = "kaggle_s3_lake.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.2"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
